pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "gmail-search"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "google-api-python-client>=2.100.0",
    "google-auth-oauthlib>=1.2.0",
    "google-genai>=1.0.0",
    "scann>=1.3.0",
    "pymupdf>=1.24.0",
    "numpy>=1.26.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.29.0",
    "click>=8.1.0",
    "tqdm>=4.66.0",
    "pyyaml>=6.0",
    "symspellpy>=6.7.0",
    "python-docx>=1.0.0",
    "openpyxl>=3.1.0",
    "icalendar>=6.0.0",
    "pillow-heif>=0.16.0",
    "dateparser>=1.2.0",
    "psutil>=5.9.0",
    # Postgres backend (paradedb + pg_search BM25). `psycopg[binary]`
    # pins the C-impl wheel so no local libpq-dev needed.
    "psycopg[binary,pool]>=3.1.18",
    "python-pptx>=1.0.0",
    "docx2txt>=0.8",  # fallback if LibreOffice not installed
    "xlrd>=2.0.0",  # legacy .xls (xlsx is openpyxl above)
    "gdown>=6.0.0",  # Drive URL parsing (gmail/drive.py)
    "beautifulsoup4>=4.12.0",
    "crawl4ai>=0.3",  # Headless URL crawler for gmail/url_fetcher.py
    # Floors bumped after `pip-audit` surfaced patchable CVEs in the
    # anaconda baseline (urllib3 streaming/redirect family, requests
    # extract_zipped_paths, nltk downloader / wordnet_app). These are
    # transitive deps; we don't import them directly, but crawl4ai
    # does for model/NLTK-data downloads at setup time.
    "urllib3>=2.5.0",
    "requests>=2.33.0",
    "nltk>=3.9.4",
]

[project.optional-dependencies]
dev = ["pytest", "requests"]

[project.scripts]
gmail-search = "gmail_search.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
markers = [
    "integration: requires live MCP+claudebox+postgres+gmail-search stack; opt-in via pytest -m integration",
]
addopts = "-m 'not integration'"
